+++ e2e/wdio/core/fixtures/appData/fixed-card-layout-content.ts
@@ -1,8 +1,7 @@
-export default {
-    cardSpacingAttr: 'margin-bottom',
-    cardSpacingValue: '16px',
-    cardWidthAttr: 'min-width',
-    cardMinWidth: '320px',
-    placeholderBorderAttr: 'border-bottom-style',
-    placeholderBorderStyle: 'dashed'
-};
+export const cardSpacingAttr = 'margin-bottom';
+export const cardSpacingValue = '16px';
+export const cardWidthAttr = 'min-width';
+export const cardMinWidth = '320px';
+export const placeholderBorderAttr = 'border-bottom-style';
+export const placeholderBorderStyle = 'dashed';
+