+++ e2e/wdio/core/tests/card.e2e-spec.ts
@@ -9,7 +9,9 @@ import {
     getElementArrayLength
 } from '../../driver/wdio';
 import { checkAttributeValueTrue, checkElArrIsClickable, checkElementText, checkElementTextValue } from '../../helper/assertion-helper';
-import CardData from '../fixtures/appData/card-content';
+import {cardTitleArr, compactAttr, analyticsTitle, badgeText, barChartCounter, barChartItems, barChartTitle, btnText,
+cardListItemText, cardSubtitle, cardTypeAttr, colorAttr, compactFont, fontSizeAttr, loaderAttr, statusColors,
+tableCardCountries, tableCardNames, tableCardPrices, tableCardStatuses, tableCardTitle, tableContentsText, tableHeaderText} from '../fixtures/appData/card-content';
 
 describe('Card test suite:', function() {
     const cardPage = new CardPo();
@@ -21,33 +23,33 @@ describe('Card test suite:', function() {
     describe('Standard card examples:', function() {
         it('should check card header', () => {
             checkElArrIsClickable(cardPage.cardHeader);
-            checkElementTextValue(cardPage.cardTitle, CardData.cardTitleArr);
+            checkElementTextValue(cardPage.cardTitle, cardTitleArr);
             expect(isElementDisplayed(cardPage.cardAvatar)).toBe(true);
             expect(isElementDisplayed(cardPage.cardCounter)).toBe(true);
             expect(isElementDisplayed(cardPage.cardSubtitle)).toBe(true);
             expect(isElementDisplayed(cardPage.cardBadge)).toBe(true);
-            expect(getText(cardPage.cardSubtitle)).toBe(CardData.cardSubtitle);
-            expect(getText(cardPage.cardBadge)).toBe(CardData.badgeText);
+            expect(getText(cardPage.cardSubtitle)).toBe(cardSubtitle);
+            expect(getText(cardPage.cardBadge)).toBe(badgeText);
         });
 
         it('should check card content', () => {
             checkElArrIsClickable(cardPage.cardListItems);
             checkElementText(cardPage.cardListItems);
-            expect(getAttributeByName(cardPage.cardAttr, CardData.cardTypeAttr, 2)).toBe('list');
+            expect(getAttributeByName(cardPage.cardAttr, cardTypeAttr, 2)).toBe('list');
         });
     });
 
     describe('Compact examples:', function() {
         it('should check its compact and do basic checks', () => {
-            checkAttributeValueTrue(cardPage.compactCardAttr, CardData.compactAttr);
+            checkAttributeValueTrue(cardPage.compactCardAttr, compactAttr);
             checkElementText(cardPage.compactCardHeader);
-            checkElementTextValue(cardPage.compactCardListItems, CardData.cardListItemText);
+            checkElementTextValue(cardPage.compactCardListItems, cardListItemText);
             checkElArrIsClickable(cardPage.compactCardHeader);
             checkElArrIsClickable(cardPage.compactCardListItems);
             // skip IE due to https://github.com/SAP/fundamental-ngx/issues/4310
             if (!browserIsIE()) {
-                expect(getCSSPropertyByName(cardPage.compactCardListItems, CardData.fontSizeAttr).value)
-                    .toBe(CardData.compactFont);
+                expect(getCSSPropertyByName(cardPage.compactCardListItems, fontSizeAttr).value)
+                    .toBe(compactFont);
                 return;
             }
             console.log('skip IE due to issue #4310');
@@ -56,7 +58,7 @@ describe('Card test suite:', function() {
 
     describe('Card loader examples:', function() {
         it('should check loading icon and attributes', () => {
-            checkAttributeValueTrue(cardPage.loaderCardAttr, CardData.loaderAttr);
+            checkAttributeValueTrue(cardPage.loaderCardAttr, loaderAttr);
             expect(isElementDisplayed(cardPage.loaderIcon)).toBe(true);
         });
     });
@@ -65,14 +67,14 @@ describe('Card test suite:', function() {
         it('should do basic checks', () => {
             checkElArrIsClickable(cardPage.ftCardHeader);
             checkElArrIsClickable(cardPage.ftCardListItems);
-            checkElementTextValue(cardPage.ftCardListItems, CardData.cardListItemText);
-            expect(getText(cardPage.ftCardHeader)).toBe(CardData.cardTitleArr[0]);
+            checkElementTextValue(cardPage.ftCardListItems, cardListItemText);
+            expect(getText(cardPage.ftCardHeader)).toBe(cardTitleArr[0]);
         });
 
         it('should check footer', () => {
             expect(isElementDisplayed(cardPage.ftFooter)).toBe(true);
             checkElArrIsClickable(cardPage.ftButtons);
-            checkElementTextValue(cardPage.ftButtons, CardData.btnText);
+            checkElementTextValue(cardPage.ftButtons, btnText);
         });
     });
 
@@ -80,7 +82,7 @@ describe('Card test suite:', function() {
         it('should check header', () => {
             checkElArrIsClickable(cardPage.kpiAnalyticsHeader);
             checkElementText(cardPage.kpiCardHeader);
-            expect(getText(cardPage.kpiCardTitle)).toBe(CardData.analyticsTitle);
+            expect(getText(cardPage.kpiCardTitle)).toBe(analyticsTitle);
             checkElementText(cardPage.kpiAnalyticsHeaderIcons);
             checkElementText(cardPage.kpiHeaderSubtitle);
         });
@@ -94,23 +96,23 @@ describe('Card test suite:', function() {
     describe('Table card examples:', function() {
         it('should check header', () => {
             checkElArrIsClickable(cardPage.tableCardHeader);
-            expect(getText(cardPage.tableCardHeader)).toBe(CardData.tableCardTitle);
+            expect(getText(cardPage.tableCardHeader)).toBe(tableCardTitle);
         });
 
         it('should check table content', () => {
-            checkElementTextValue(cardPage.tableCardTableHeader, CardData.tableHeaderText);
-            checkElementTextValue(cardPage.tableCardItemNames, CardData.tableCardNames);
-            checkElementTextValue(cardPage.tableCardItemCountries, CardData.tableCardCountries);
-            checkElementTextValue(cardPage.tableCardItemPrices, CardData.tableCardPrices);
-            checkElementTextValue(cardPage.tableCardItemStatuses, CardData.tableCardStatuses);
+            checkElementTextValue(cardPage.tableCardTableHeader, tableHeaderText);
+            checkElementTextValue(cardPage.tableCardItemNames, tableCardNames);
+            checkElementTextValue(cardPage.tableCardItemCountries, tableCardCountries);
+            checkElementTextValue(cardPage.tableCardItemPrices, tableCardPrices);
+            checkElementTextValue(cardPage.tableCardItemStatuses, tableCardStatuses);
             checkElArrIsClickable(cardPage.tableCardItems);
         });
 
         it('should check status colors', () => {
             const statusesCount = getElementArrayLength(cardPage.tableCardItemStatuses);
             for (let i = 0; statusesCount > i; i++) {
-                expect(getCSSPropertyByName(cardPage.tableCardItemStatuses, CardData.colorAttr, i).value)
-                    .toContain(CardData.statusColors[i]);
+                expect(getCSSPropertyByName(cardPage.tableCardItemStatuses, colorAttr, i).value)
+                    .toContain(statusColors[i]);
             }
         });
     });
@@ -118,12 +120,12 @@ describe('Card test suite:', function() {
     describe('Bar chart list examples:', function() {
         it('should check header', () => {
             checkElArrIsClickable(cardPage.barChartHeader);
-            expect(getText(cardPage.barChartTitle)).toBe(CardData.barChartTitle);
-            expect(getText(cardPage.barChartCounter)).toBe(CardData.barChartCounter);
+            expect(getText(cardPage.barChartTitle)).toBe(barChartTitle);
+            expect(getText(cardPage.barChartCounter)).toBe(barChartCounter);
         });
 
         it('should check chart content', () => {
-            checkElementTextValue(cardPage.barChartItems, CardData.barChartItems);
+            checkElementTextValue(cardPage.barChartItems, barChartItems);
             expect(isElementDisplayed(cardPage.barCharBars)).toBe(true);
         });
     });