+++ e2e/wdio/platform/fixtures/testData/input-group.ts
@@ -1,5 +1,3 @@
-export default {
-    numeric_value: '1234',
-    string_value: 'test',
-    email_value: 'test@test.com'
-}
+export const numeric_value = '1234';
+export const string_value = 'test';
+export const email_value = 'test@test.com';