+++ CHANGELOG.md
@@ -2,6 +2,55 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.27.1-rc.2](https://github.com/SAP/fundamental-ngx/compare/v0.27.1-rc.1...v0.27.1-rc.2) (2021-02-16)
+
+### [0.27.1-rc.1](https://github.com/SAP/fundamental-ngx/compare/v0.27.1-rc.0...v0.27.1-rc.1) (2021-02-16)
+
+### [0.27.1-rc.0](https://github.com/SAP/fundamental-ngx/compare/v0.27.0...v0.27.1-rc.0) (2021-02-15)
+
+## [0.27.0](https://github.com/SAP/fundamental-ngx/compare/v0.27.0-rc.5...v0.27.0) (2021-02-12)
+
+## [0.27.0-rc.5](https://github.com/SAP/fundamental-ngx/compare/v0.27.0-rc.4...v0.27.0-rc.5) (2021-02-12)
+
+
+### Bug Fixes
+
+* Readonly select and arrow down fix ([#4631](https://github.com/SAP/fundamental-ngx/issues/4631)) ([a92d96e](https://github.com/SAP/fundamental-ngx/commit/a92d96ef0d8da1eff5f64cc94e8bb5e082ced395))
+
+## [0.27.0-rc.4](https://github.com/SAP/fundamental-ngx/compare/v0.27.0-rc.3...v0.27.0-rc.4) (2021-02-12)
+
+## [0.27.0-rc.3](https://github.com/SAP/fundamental-ngx/compare/v0.27.0-rc.2...v0.27.0-rc.3) (2021-02-11)
+
+
+### Bug Fixes
+
+* improvements to the wizard in dialog ([#4580](https://github.com/SAP/fundamental-ngx/issues/4580)) ([faf00ce](https://github.com/SAP/fundamental-ngx/commit/faf00cef2a617d81bd3c8384829992c25515e8a4))
+
+## [0.27.0-rc.2](https://github.com/SAP/fundamental-ngx/compare/v0.27.0-rc.1...v0.27.0-rc.2) (2021-02-11)
+
+
+### Bug Fixes
+
+* (Core) Add token validate function to MultiInput ([#4621](https://github.com/SAP/fundamental-ngx/issues/4621)) ([d55fc97](https://github.com/SAP/fundamental-ngx/commit/d55fc97c6a6ddff59efa9e9fed38d93d52d19ce7))
+
+## [0.27.0-rc.1](https://github.com/SAP/fundamental-ngx/compare/v0.27.0-rc.0...v0.27.0-rc.1) (2021-02-11)
+
+
+### Bug Fixes
+
+* (Platform) Empty table filtering options ([0f695c9](https://github.com/SAP/fundamental-ngx/commit/0f695c9899dcf9100e714195c0274abdc5200b1a))
+
+## [0.27.0-rc.0](https://github.com/SAP/fundamental-ngx/compare/v0.26.2-rc.34...v0.27.0-rc.0) (2021-02-11)
+
+
+### ⚠ BREAKING CHANGES
+
+* The avatar in card header needs to set size
+
+### Bug Fixes
+
+* expose avatar inputs as part of card header ([#4638](https://github.com/SAP/fundamental-ngx/issues/4638)) ([29a31ea](https://github.com/SAP/fundamental-ngx/commit/29a31eaa61bb273340d2ed5d85458a200c75ee56))
+
 ### [0.26.2-rc.34](https://github.com/SAP/fundamental-ngx/compare/v0.26.2-rc.33...v0.26.2-rc.34) (2021-02-11)